chore: extract repeated helper `createFilesList` into a common module
